min() cast
git-svn-id: svn+ssh://beta.visl.sdu.dk/usr/local/svn/repos/visl/tools/vislcg3/trunk@12114 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/GrammarApplicator_runRules.cpp
@@ -1716,7 +1716,7 @@ uint32_t GrammarApplicator::runRulesOnSingleWindow(SingleWindow& current, const 
 								else if (type == K_MOVE_AFTER) {
 									spot = edges.back()->local_number + 1;
 								}
-								spot = std::min(spot, current.cohorts.size());
+								spot = std::min(spot, static_cast<uint32_t>(current.cohorts.size()));
 
 								while (!cohorts.empty()) {
 									foreach (iter, cohorts.back()->readings) {
